Move extractHighlights to separate src folder
