+++ src/extractHighlight.ts
